+++ src/main/resources/db/migration/V3__Create_question_table.sql
@@ -4,7 +4,7 @@ create table question
 	description text,
 	gmt_create bigint,
 	gmt_modified bigint,
-	creator int,
+	creator bigint,
 	comment_count int default 0,
 	view_count int default 0,
 	like_count int default 0,
