+++ src/main/resources/db/migration/V4__Add_id_col_to_question_table.sql
@@ -1,5 +1,5 @@
 alter table QUESTION
-	add id int auto_increment;
+	add id bigint auto_increment;
 
 alter table QUESTION
 	add constraint QUESTION_pk
